Time Files Creation and Editing Funtion Added
The Movie File Editer class contains all the goodness that opens a GUI for the user to make new time files for movies and edit existing files. There is a sweet JOptionPane that asks to confirm if you want to save. It is simply a safeguard against accidently messing up a file.
I added something to the YeltsinScheduler class (sorry to go into that class, Johnsky). I just added a method to get the current time of the movie so that I could add the "get that there current time" function to the times file editing GUI.
In the main Yeltsin class, I added some buttons to the central GUI for opening a frame for editing the time files. I also made the name of the file selected through the moveMenu global variable.